Refactor slice folder names
