+++ src/components/Cart/CountModifier/index.tsx
@@ -2,9 +2,9 @@ import debounce from "lodash.debounce";
 import React, { type FC } from "react";
 import { useTypedDispatch, useTypedSelector } from "@/hooks/redux";
 import { cartSelector } from "@/store/selectors/cart";
-import { modifyCartItemCount, removeCartItem } from "@/store/slices/cartSlice/cart-thunks";
-import { type CounterToChange } from "@/store/slices/cartSlice/helpers";
-import { CountAction } from "@/store/slices/cartSlice/Types";
+import { modifyCartItemCount, removeCartItem } from "@/store/slices/cart/cart-thunks";
+import { type CounterToChange } from "@/store/slices/cart/helpers";
+import { CountAction } from "@/store/slices/cart/Types";
 import CounterInput from "@/features/CounterInput/CounterInput";
 import DecreaseCounter from "@/features/CounterInput/DecreaseCounter";
 import IncreaseCounter from "@/features/CounterInput/IncreaseCounter";
