+++ src/components/Cart/index.tsx
@@ -2,8 +2,8 @@ import { Modal, Paper } from "@mui/material";
 import React, { type FC } from "react";
 import { useTypedDispatch, useTypedSelector } from "../../hooks/redux";
 import { cartSelector } from "../../store/selectors/cart";
-import { getCartItems } from "../../store/slices/cartSlice/cart-thunks";
-import { RequestStatus } from "../../store/slices/cartSlice/Types";
+import { getCartItems } from "../../store/slices/cart/cart-thunks";
+import { RequestStatus } from "../../store/slices/cart/Types";
 import HeaderWithClose from "../common/HeaderWithClose";
 import BasicPreloader from "../Loaders/BasicPreloader";
 import CartItemLoader from "../Loaders/CartItemLoader";
