+++ src/components/Search/index.tsx
@@ -2,7 +2,7 @@ import { Container, Paper } from "@mui/material";
 import React from "react";
 import { useTypedSelector } from "../../hooks/redux";
 import { getSearchSlice } from "@/store/selectors/search";
-import { setIsSearchActive } from "@/store/slices/searchSlice";
+import { setIsSearchActive } from "@/store/slices/search";
 import HeaderWithClose from "@/components/common/HeaderWithClose";
 import ProductsSearch from "./ProductsSearch";
 import SearchResults from "./SearchResults";
