+++ src/features/CounterInput/DecreaseCounter.tsx
@@ -1,6 +1,6 @@
 import IndeterminateCheckBoxIcon from "@mui/icons-material/IndeterminateCheckBox";
 import React, { type FC } from "react";
-import { CountAction } from "../../store/slices/cartSlice/Types";
+import { CountAction } from "../../store/slices/cart/Types";
 
 interface Props {
     modifyCounter: (countAction: CountAction) => void;
