+++ src/features/ProductSizes/index.tsx
@@ -1,11 +1,11 @@
-import { type SelectedCartItem } from "@/store/slices/cartSlice/Types";
+import { type SelectedCartItem } from "@/store/slices/cart/Types";
 import { Stack, Typography } from "@mui/material";
 import classNames from "classnames";
 import React, { type FC } from "react";
 import { useGetSizeItems } from "../../components/ProductCard/useGetSizeItems";
 import { useTypedDispatch } from "../../hooks/redux";
 import { type IProductCard } from "../../types/IProductCard";
-import { addToCart } from "../../store/slices/cartSlice/cart-thunks";
+import { addToCart } from "../../store/slices/cart/cart-thunks";
 import ToggleFavorite from "../ToggleFavorite";
 import AddToCartButton, { type ColorVariant } from "./AddToCartButton";
