+++ src/store/selectors/cart.ts
@@ -1,4 +1,4 @@
-import { type CartInitialState } from "../slices/cartSlice";
+import { type CartInitialState } from "../slices/cart";
 import { type RootState } from "../store";
 
 export const cartSelector = (state: RootState): CartInitialState => state.cartSlice;
