+++ src/store/slices/cart/helpers.ts
