+++ src/store/slices/search/index.ts
